+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Renamed `*Plume` classes to `*P`, for brevity; for example, use `CollectionsP` instead of `CollectionsPlume`.
 - Removed all deprecated classes and mehods.
 
-## 1.9.1 (????-??-??)
+## 1.9.1 (2024-04-06)
 
 - `FilesPlume`:
     * deprecate `readFile()` in favor of `fileContents()`